The hp bar ux is now smooth

+++ Assets/Scripts/Battle/BattleStates/PlayTurn.cs
@@ -55,7 +55,7 @@ public class PlayTurn : State
         var endHp = defender.GetHp();
         SoundManager.__instance__.PlaySoundEffect(BattleUI.GetSound("DamageSound"));
         // TODO : Maybe use a factory
-        Action<ITween<float>> tweenHpCallback = (t) => { defender.UpdateHpBar((int)t.CurrentValue); };
+        Action<ITween<float>> tweenHpCallback = (t) => { defender.UpdateHpBar(t.CurrentValue); };
         defender.gameObject.Tween("hpDefend", startHp, endHp, 0.4f, TweenScaleFunctions.SineEaseIn, tweenHpCallback, (t) =>
         {
             StateStack.Pop();
